server/requirements.txt
asgiref==3.6.0
django-heroku==0.3.1
Django==4.1.6
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.4.0
virtualenv==20.16.7
dj-database-url==1.0.0